akinator-ai/requirements.txt
# Core LangChain & LangGraph
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.2.0
langchain-core>=0.3.0

# LangSmith for monitoring and learning
langsmith>=0.1.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# CLI Enhancement
rich>=13.0.0
click>=8.0.0

# Data Processing
typing-extensions>=4.0.0

# Optional Web UI (Phase 9)
# streamlit>=1.28.0
# gradio>=4.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
